requirements.txt
CurrencyConverter>=0.14.4
emoji>=0.6.0
GitPython>=3.1.11
googletrans>=3.0.0
gTTS>=2.2.1
pydub>=0.24.1
python-dateutil>=2.8.1
python-dotenv>=0.15.0
requests>=2.25.1
SpeechRecognition>=3.8.1
speedtest-cli>=2.1.2
Telethon>=1.18.2
uptime>=3.0.1
urllib3>=1.26.2

